requirements.txt
streamlit
streamlit-lottie
ultralytics
opencv-python
pyttsx3
SpeechRecognition
transformers
torch
torchvision
Pillow
requests
